Move MessageSource/Sink traits to mod builders
- These two traits are related to message box builders
  and not directly to message boxes.
- Similarly, struct RequestResponseHandler has been moved to mod
  message_boxes.
