+++ crates/core/tedge_actors/src/lib.rs
@@ -375,7 +375,6 @@ mod actors;
 pub mod builders;
 mod channels;
 mod errors;
-mod handles;
 mod keyed_messages;
 pub mod message_boxes;
 mod messages;
@@ -389,7 +388,6 @@ pub use actors::*;
 pub use builders::*;
 pub use channels::*;
 pub use errors::*;
-pub use handles::*;
 pub use keyed_messages::*;
 pub use message_boxes::*;
 pub use messages::*;
